+++ src/main/kotlin/nexters/admin/domain/user/Password.kt
@@ -1,4 +1,4 @@
-package nexters.admin.domain.member
+package nexters.admin.domain.user
 
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.core.JsonParser
