+++ src/main/kotlin/nexters/admin/repository/MemberRepository.kt
@@ -1,6 +1,6 @@
 package nexters.admin.repository
 
-import nexters.admin.domain.member.Member
+import nexters.admin.domain.user.member.Member
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface MemberRepository : JpaRepository<Member, Long>
