+++ src/test/kotlin/nexters/admin/domain/user/PasswordTest.kt
@@ -1,4 +1,4 @@
-package nexters.admin.domain.member
+package nexters.admin.domain.user
 
 import io.kotest.matchers.shouldNotBe
 import org.junit.jupiter.api.Test
